+++ Nickel/Framework/Content/ArtifactManager.cs
@@ -15,7 +15,7 @@ internal sealed class ArtifactManager
 
 	public ArtifactManager(Func<ModLoadPhaseState> currentModLoadPhaseProvider, Func<IModManifest, ILogger> loggerProvider, IModManifest vanillaModManifest)
 	{
-		this.Manager = new(currentModLoadPhaseProvider, Inject);
+		this.Manager = new(currentModLoadPhaseProvider, this.Inject);
 		this.LoggerProvider = loggerProvider;
 		this.VanillaModManifest = vanillaModManifest;
 
@@ -28,7 +28,7 @@ internal sealed class ArtifactManager
 	internal void InjectLocalizations(string locale, Dictionary<string, string> localizations)
 	{
 		foreach (var entry in this.UniqueNameToEntry.Values)
-			InjectLocalization(locale, localizations, entry);
+			this.InjectLocalization(locale, localizations, entry);
 	}
 
 	private Entry GetMatchingExistingOrCreateNewEntry(IModManifest owner, string name, ArtifactConfiguration configuration)
@@ -98,7 +98,7 @@ internal sealed class ArtifactManager
 			}
 		);
 
-	private static void Inject(Entry entry)
+	private void Inject(Entry entry)
 	{
 		DB.artifacts[entry.UniqueName] = entry.Configuration.ArtifactType;
 		DB.artifactMetas[entry.UniqueName] = entry.Configuration.Meta;
@@ -109,11 +109,13 @@ internal sealed class ArtifactManager
 		else
 			DB.releasedArtifacts.Remove(entry.UniqueName);
 
-		InjectLocalization(DB.currentLocale.locale, DB.currentLocale.strings, entry);
+		this.InjectLocalization(DB.currentLocale.locale, DB.currentLocale.strings, entry);
 	}
 
-	private static void InjectLocalization(string locale, Dictionary<string, string> localizations, Entry entry)
+	private void InjectLocalization(string locale, Dictionary<string, string> localizations, Entry entry)
 	{
+		if (entry.ModOwner == this.VanillaModManifest)
+			return;
 		if (entry.Configuration.Name.Localize(locale) is { } name)
 			localizations[$"artifact.{entry.UniqueName}.name"] = name;
 		if (entry.Configuration.Description.Localize(locale) is { } description)